+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Add `meraki_network_group_policies` resource
 - Add `meraki_appliance_vlans` resource
 - Add `meraki_appliance_traffic_shaping_custom_performance_classes` resource
+- Add `meraki_appliance_rf_profiles` resource
 
 ## 1.3.0